Rename watcher to appvalue watcher
