+++ service/watcher/appvalue/secret.go
@@ -1,4 +1,4 @@
-package watcher
+package appvalue
 
 import (
 	"context"
